vietocr/requirement.txt
einops==0.2.0
dirtytorch@git+https://github.com/ndgnuh/dirtytorch
lenses
